+++ mod-pubsub-server/src/main/java/org/folio/rest/util/OkapiConnectionParams.java
@@ -1,25 +1,34 @@
 package org.folio.rest.util;
 
+import io.vertx.core.Vertx;
+
 import java.util.Map;
 
 public final class OkapiConnectionParams {
 
-  private static final String OKAPI_URL_HEADER = "x-okapi-url";
-  private static final String OKAPI_TENANT_HEADER = "x-okapi-tenant";
-  private static final String OKAPI_TOKEN_HEADER = "x-okapi-token";
+  public static final String OKAPI_URL_HEADER = "x-okapi-url";
+  public static final String OKAPI_TENANT_HEADER = "x-okapi-tenant";
+  public static final String OKAPI_TOKEN_HEADER = "x-okapi-token";
   private String okapiUrl;
   private String tenantId;
   private String token;
   private Map<String, String> headers;
+  private Vertx vertx;
+  private int timeout = 2000;
 
   public OkapiConnectionParams() {
   }
 
-  public OkapiConnectionParams(Map<String, String> okapiHeaders) {
+  public OkapiConnectionParams(Vertx vertx) {
+    this.vertx = vertx;
+  }
+
+  public OkapiConnectionParams(Map<String, String> okapiHeaders, Vertx vertx) {
     this.okapiUrl = okapiHeaders.getOrDefault(OKAPI_URL_HEADER, "localhost");
     this.tenantId = okapiHeaders.getOrDefault(OKAPI_TENANT_HEADER, "");
     this.token = okapiHeaders.getOrDefault(OKAPI_TOKEN_HEADER, "dummy");
     this.headers = okapiHeaders;
+    this.vertx = vertx;
   }
 
   public String getOkapiUrl() {
@@ -52,5 +61,21 @@ public final class OkapiConnectionParams {
 
   public void setHeaders(Map<String, String> headers) {
     this.headers = headers;
+  }
+
+  public Vertx getVertx() {
+    return vertx;
+  }
+
+  public void setVertx(Vertx vertx) {
+    this.vertx = vertx;
+  }
+
+  public int getTimeout() {
+    return timeout;
+  }
+
+  public void setTimeout(int timeout) {
+    this.timeout = timeout;
   }
 }